+++ finagle-core/src/main/scala/com/twitter/finagle/Stack.scala
@@ -622,6 +622,46 @@ object Stack {
         next
       )
   }
+
+  /** A module of 7 parameters. */
+  abstract class Module7[
+    P1: Param,
+    P2: Param,
+    P3: Param,
+    P4: Param,
+    P5: Param,
+    P6: Param,
+    P7: Param,
+    T] extends Stackable[T] {
+    final val parameters: Seq[Stack.Param[_]] = Seq(
+      implicitly[Param[P1]],
+      implicitly[Param[P2]],
+      implicitly[Param[P3]],
+      implicitly[Param[P4]],
+      implicitly[Param[P5]],
+      implicitly[Param[P6]],
+      implicitly[Param[P7]]
+    )
+    def make(p1: P1, p2: P2, p3: P3, p4: P4, p5: P5, p6: P6, p7: P7, next: T): T
+    def toStack(next: Stack[T]): Stack[T] =
+      Node(
+        this,
+        (prms, next) =>
+          Leaf(
+            this,
+            make(
+              prms[P1],
+              prms[P2],
+              prms[P3],
+              prms[P4],
+              prms[P5],
+              prms[P6],
+              prms[P7],
+              next.make(prms))
+        ),
+        next
+      )
+  }
 }
 
 /**